Refactor Pokemon.py to import csv module and add ability_list attribute
The code changes in Pokemon.py refactor the code to import the csv module and add the ability_list attribute to the Pokemon class. This change allows for reading and storing the abilities of each Pokemon from a CSV file, improving the flexibility and maintainability of the code.

This commit message follows the convention of starting with a verb (Refactor) to indicate the type of change, followed by a concise and descriptive message of the change.